+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# Version 1.2.2
+# Version 1.3
+
+## Converters
+- Add support for AMLTK.
 
 ## Ablation Paths
 - Added ablation paths as a plugin.
@@ -9,8 +12,23 @@
 # Version 1.2.1
 
 ## Quality of Life
-- Runs now get displayed with their parent directory for better distinguishability
+- Runs now get displayed with their parent directory for better distinguishability.
 - Increase plot font sizes.
+- Add a simple loading bar functionality for longer runs.
+
+## General
+- Seed is now required in the Recorder.
+
+## Bug-Fixes
+- Use normalized LPI importance via variance instead of importance over mean (#152)
+- Return nan as importance values if variance is 0. for a hyperparameter / budget (#152)
+
+## Plugins
+- Show a run's hoover-text for the actual budget of a trial in Cost over Time with Combined budget (#154).
+- Use highest budget as default budget for Cost over Time instead of Combined.
+- Show best value / config for each objective instead of merged objective in Overview (#159).
+- Use chosen objective instead of merged objective to get the incumbent for the calculation of LPI importance (#159).
+- Add total runtime in overview (#155).
 
 # Version 1.2